requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.8.0rc2
diff-match-patch==20241021
Django==5.1.5
django-ckeditor-5==0.2.15
django-filter==24.3
django-import-export==4.3.4
django-js-asset==3.0.1
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
idna==3.10
oauthlib==3.2.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
tablib==3.8.0
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
user-agents==2.2.0
validators==0.34.0
